+++ packages/ui-core/src/components/TextField/TextField.tsx
@@ -1,15 +1,23 @@
 import * as React from 'react';
 import { useCallback, useEffect, useState, useRef, useMemo } from 'react';
-import { cnCreate, detectTouch, filterDataAttrs } from '@megafon/ui-helpers';
+import { cnCreate, filterDataAttrs } from '@megafon/ui-helpers';
 import Hide from '@megafon/ui-icons/basic-24-hide_24.svg';
 import Show from '@megafon/ui-icons/basic-24-show_24.svg';
-import ErrorIcon from '@megafon/ui-icons/system-24-cancel_24.svg';
+import ClearIcon from '@megafon/ui-icons/system-24-cancel_24.svg';
 import CheckedIcon from '@megafon/ui-icons/system-24-checked_24.svg';
+import throttle from 'lodash.throttle';
 import * as PropTypes from 'prop-types';
 import InputMask from 'react-input-mask';
-import InputLabel from '../InputLabel/InputLabel';
-import Paragraph from '../Paragraph/Paragraph';
+import throttleTime from 'constants/throttleTime';
+import ResizeIcon from './i/textarea-resizer.svg';
 import './TextField.less';
+
+const DEFAULT_PLACEHOLDERS = {
+    email: 'E-mail',
+    tel: 'Номер телефона',
+    password: 'Пароль',
+    text: 'Текст',
+};
 
 export const Verification = {
     VALID: 'valid',
@@ -54,6 +62,8 @@ export type TextFieldProps = {
     inputRef?: (node: HTMLInputElement | HTMLTextAreaElement) => void;
     /** Имя поля */
     name?: string;
+    /** Запрещает отображение плейсхолдера */
+    hidePlaceholder?: boolean;
     /** Отображаемый текст при отсутствии значения */
     placeholder?: string;
     /** Атрибут корневого DOM элемента компонента */
@@ -74,7 +84,9 @@ export type TextFieldProps = {
     /** Дополнительный класс корневого элемента */
     className?: string;
     /** Дополнительные классы элементов */
-    classes?: { input?: string };
+    classes?: {
+        input?: string;
+    };
     /** Дополнительные data атрибуты к внутренним элементам */
     dataAttrs?: {
         root?: Record<string, string>;
@@ -103,10 +115,11 @@ export type TextFieldProps = {
     onCustomIconClick?: (e: React.MouseEvent<HTMLDivElement>) => void;
 };
 
-const TEXTAREA_MIN_HEIGHT = 96;
-const TEXTAREA_MAX_HEIGHT = 168;
-const ROW_HEIGHT = 24;
+const TEXTAREA_MIN_HEIGHT = 48;
+const TEXTAREA_MAX_HEIGHT = 144;
+const DEFAULT_LABEL_TOP_POSITION = 16;
 const DEFAULT_ROW_COUNT = 3;
+const ROW_HEIGHT = 24;
 
 const cn = cnCreate('mfui-text-field');
 const TextField: React.FC<TextFieldProps> = ({
@@ -122,6 +135,7 @@ const TextField: React.FC<TextFieldProps> = ({
     symbolCounter,
     textarea = false,
     name,
+    hidePlaceholder = false,
     placeholder,
     required,
     isControlled = false,
@@ -135,7 +149,7 @@ const TextField: React.FC<TextFieldProps> = ({
     type = 'text',
     value = '',
     verification,
-    noticeText,
+    noticeText = '',
     inputRef,
     inputMode,
     autoComplete,
@@ -144,17 +158,28 @@ const TextField: React.FC<TextFieldProps> = ({
 }) => {
     const [isPasswordHidden, setPasswordHidden] = useState<boolean>(true);
     const [inputValue, setInputValue] = useState<string | number | undefined>(value);
-    const [initialTextareaHeight, setInitialTextareaHeight] = useState(TEXTAREA_MIN_HEIGHT);
-    const [isTextareaResized, setIsTextareaResized] = useState(false);
+    const [initialTextareaHeight, setInitialTextareaHeight] = useState(ROW_HEIGHT * DEFAULT_ROW_COUNT);
     const [isMaxLimitExceeded, setIsMaxLimitExceeded] = useState(false);
-    const [isTouch, setTouch] = useState(false);
+    const [currentNoticeText, setCurrentNoticeText] = useState(noticeText);
+    const [isTextareaResizeFocused, setIsTextareaResizeFocused] = useState(false);
+    const [isTextareaResized, setIsTextareaResized] = useState(false);
+
     const fieldNode = useRef<HTMLInputElement | HTMLTextAreaElement>();
+    const labelRef = useRef<HTMLLabelElement>(null);
+    const resizerRef = useRef<HTMLDivElement>(null);
 
     const isPasswordType: boolean = useMemo(() => type === 'password', [type]);
     const isVisiblePassword: boolean = useMemo(
         () => isPasswordType && !isPasswordHidden,
         [isPasswordHidden, isPasswordType],
     );
+
+    const hasValue = isControlled ? !!value : !!inputValue;
+    const isValidVerification = verification === Verification.VALID && !disabled;
+    const isErrorVerification = verification === Verification.ERROR && !disabled;
+    const hasValueForClear = hasValue && !isPasswordType && !customIcon && !isValidVerification;
+    const hasClearIcon = !disabled && (hasValueForClear || isErrorVerification);
+    const actualPlaceholder = placeholder || DEFAULT_PLACEHOLDERS[type];
 
     const checkSymbolMaxLimit = useCallback(
         (textareaValue: string | number = ''): void => {
@@ -173,13 +198,64 @@ const TextField: React.FC<TextFieldProps> = ({
     }, [value, checkSymbolMaxLimit, isControlled]);
 
     useEffect(() => {
-        setTouch(detectTouch());
-    }, []);
+        noticeText && setCurrentNoticeText(noticeText);
+    }, [noticeText]);
+
+    useEffect(() => {
+        if (!resizerRef.current || textarea !== TextareaTypes.FLEXIBLE) {
+            return;
+        }
+
+        const handleStartResize = (downEvent: MouseEvent | TouchEvent): void => {
+            if (!fieldNode.current) {
+                return;
+            }
+
+            setIsTextareaResizeFocused(true);
+
+            downEvent.preventDefault();
+
+            const originalHeight = parseFloat(
+                getComputedStyle(fieldNode.current).getPropertyValue('height').replace('px', ''),
+            );
+            const originalCoordinateY =
+                (downEvent as MouseEvent).clientY || (downEvent as TouchEvent).touches[0].clientY;
+
+            const handleResize = throttle((moveEvent: MouseEvent | TouchEvent): void => {
+                const currentCoordinateY =
+                    (moveEvent as MouseEvent).clientY || (moveEvent as TouchEvent).touches[0].clientY;
+                const resizeHeight = originalHeight + (currentCoordinateY - originalCoordinateY);
+
+                const updatedHeight = resizeHeight < TEXTAREA_MIN_HEIGHT ? TEXTAREA_MIN_HEIGHT : resizeHeight;
+
+                setInitialTextareaHeight(updatedHeight);
+                setIsTextareaResized(true);
+            }, throttleTime.resizeTextarea);
+
+            const handleResizeCancel = (): void => {
+                setIsTextareaResizeFocused(false);
+                window.removeEventListener('mousemove', handleResize);
+                window.removeEventListener('touchmove', handleResize);
+
+                window.removeEventListener('mouseup', handleResizeCancel);
+                window.removeEventListener('touchend', handleResizeCancel);
+            };
+
+            window.addEventListener('mousemove', handleResize);
+            window.addEventListener('touchmove', handleResize);
+
+            window.addEventListener('mouseup', handleResizeCancel);
+            window.addEventListener('touchend', handleResizeCancel);
+        };
+
+        resizerRef.current.addEventListener('mousedown', handleStartResize);
+        resizerRef.current.addEventListener('touchstart', handleStartResize);
+    }, [textarea]);
 
     const togglePasswordHiding = useCallback(() => setPasswordHidden(prevPassState => !prevPassState), []);
 
     const setTextareaHeight = (): void => {
-        if (!fieldNode?.current) {
+        if (!fieldNode?.current || isTextareaResized) {
             return;
         }
 
@@ -187,15 +263,11 @@ const TextField: React.FC<TextFieldProps> = ({
             current: { scrollHeight },
         } = fieldNode;
 
-        if (!isTextareaResized) {
-            const extraRowCount = Math.round((scrollHeight - TEXTAREA_MIN_HEIGHT) / ROW_HEIGHT);
-            const newHeight =
-                extraRowCount <= DEFAULT_ROW_COUNT
-                    ? TEXTAREA_MIN_HEIGHT + ROW_HEIGHT * extraRowCount
-                    : TEXTAREA_MAX_HEIGHT;
+        const extraRowCount = Math.round((scrollHeight - 28 - TEXTAREA_MIN_HEIGHT) / ROW_HEIGHT);
+        const newHeight =
+            extraRowCount <= DEFAULT_ROW_COUNT ? TEXTAREA_MIN_HEIGHT + ROW_HEIGHT * extraRowCount : TEXTAREA_MAX_HEIGHT;
 
-            setInitialTextareaHeight(newHeight);
-        }
+        setInitialTextareaHeight(newHeight);
     };
 
     const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>): void => {
@@ -209,22 +281,31 @@ const TextField: React.FC<TextFieldProps> = ({
         onChange?.(e);
     };
 
-    const handleTextareaClick = () => {
-        if (!fieldNode?.current) {
+    const handleNoticeTransitionEnd = useCallback(() => {
+        !noticeText && setCurrentNoticeText(noticeText);
+    }, [noticeText]);
+
+    const handleTextareaScroll = () => {
+        if (!fieldNode?.current || !labelRef?.current) {
             return;
         }
 
         const {
-            current: { offsetHeight: textAreaHeight },
+            current: { scrollTop },
         } = fieldNode;
 
-        setIsTextareaResized(textAreaHeight < initialTextareaHeight);
+        if (!scrollTop) {
+            labelRef.current.style.top = '';
+
+            return;
+        }
+
+        labelRef.current.style.top = `${DEFAULT_LABEL_TOP_POSITION - scrollTop}px`;
     };
 
     const handleIconClick = useCallback(
         e => {
-            const { ERROR } = Verification;
-            const isClearFuncAvailable = !customIcon && !onCustomIconClick && verification === ERROR;
+            const isClearFuncAvailable = !customIcon && !onCustomIconClick && hasClearIcon;
             const { current: field } = fieldNode;
 
             isPasswordType && togglePasswordHiding();
@@ -234,7 +315,7 @@ const TextField: React.FC<TextFieldProps> = ({
                 field?.focus();
             }
         },
-        [isPasswordType, togglePasswordHiding, onCustomIconClick, verification, customIcon, isControlled],
+        [isPasswordType, togglePasswordHiding, onCustomIconClick, customIcon, isControlled, hasClearIcon],
     );
 
     const handleFocus = useCallback(
@@ -258,7 +339,6 @@ const TextField: React.FC<TextFieldProps> = ({
     );
 
     const textareaType = textarea === TextareaTypes.FLEXIBLE ? TextareaTypes.FLEXIBLE : TextareaTypes.FIXED;
-    const hasScrolling = initialTextareaHeight >= TEXTAREA_MAX_HEIGHT || isTextareaResized;
 
     const commonParams = {
         ...filterDataAttrs(dataAttrs?.input),
@@ -271,7 +351,7 @@ const TextField: React.FC<TextFieldProps> = ({
         onFocus: handleFocus,
         onKeyUp,
         maxLength,
-        placeholder,
+        placeholder: hidePlaceholder ? ' ' : actualPlaceholder,
         required,
         inputMode,
     };
@@ -295,7 +375,8 @@ const TextField: React.FC<TextFieldProps> = ({
             'field',
             {
                 type: textareaType,
-                scrolling: hasScrolling,
+                textarea,
+                resized: isTextareaResizeFocused,
             },
             input,
         ),
@@ -310,61 +391,76 @@ const TextField: React.FC<TextFieldProps> = ({
         inputRef?.(node);
     };
 
-    const getIcon = (): React.ReactNode | null => {
+    const getIcon = (): JSX.Element | undefined => {
         switch (true) {
             case !!customIcon:
                 return customIcon;
-            case verification === Verification.ERROR:
-                return <ErrorIcon className={cn('icon')} />;
-            case verification === Verification.VALID:
-                return <CheckedIcon className={cn('icon')} />;
             case isPasswordType && isPasswordHidden:
                 return <Hide className={cn('icon')} />;
             case isPasswordType && !isPasswordHidden:
                 return <Show className={cn('icon')} />;
+            case isValidVerification:
+                return <CheckedIcon className={cn('icon')} />;
+            case hasClearIcon:
+                return <ClearIcon className={cn('icon', { clear: true })} />;
             default:
-                return null;
+                return undefined;
         }
     };
 
-    const renderTextarea = (): React.ReactNode => (
-        <>
-            <textarea
-                {...textareaParams}
-                onClick={handleTextareaClick}
-                style={{ height: `${initialTextareaHeight}px` }}
-                ref={getFieldNode}
-            />
-        </>
-    );
-
-    const renderIconBlock = () => {
-        const icon: React.ReactNode | null = getIcon();
+    const renderLabel = (): JSX.Element => {
+        const currentLabel = label || actualPlaceholder;
 
         return (
-            icon && (
-                <div
-                    {...filterDataAttrs(dataAttrs?.iconBox)}
-                    className={cn('icon-box', {
-                        error: verification === Verification.ERROR && !customIcon,
-                        password: isPasswordType,
-                        'custom-handler': !!onCustomIconClick,
-                    })}
-                    onClick={handleIconClick}
-                >
-                    {icon}
-                </div>
-            )
+            <label {...filterDataAttrs(dataAttrs?.label)} htmlFor={id} className={cn('label')} ref={labelRef}>
+                {currentLabel}
+                {required && <span className={cn('require-mark')}>*</span>}
+            </label>
         );
     };
 
-    const renderInput = (): React.ReactNode => (
+    const renderTextarea = (): JSX.Element => (
+        <>
+            <textarea
+                {...textareaParams}
+                onScroll={handleTextareaScroll}
+                style={{ height: `${initialTextareaHeight}px` }}
+                ref={getFieldNode}
+            />
+            {renderLabel()}
+        </>
+    );
+
+    const renderIconBlock = (): JSX.Element | undefined => {
+        const icon: React.ReactNode | null = getIcon();
+
+        if (!icon) {
+            return undefined;
+        }
+
+        return (
+            <div
+                {...filterDataAttrs(dataAttrs?.iconBox)}
+                className={cn('icon-box', {
+                    error: isErrorVerification && !customIcon,
+                    password: isPasswordType,
+                    'custom-handler': !!onCustomIconClick,
+                })}
+                onClick={handleIconClick}
+            >
+                {icon}
+            </div>
+        );
+    };
+
+    const renderInput = (): JSX.Element => (
         <>
             {mask ? (
                 <InputMask {...inputParams} {...inputMaskParams} inputRef={getFieldNode} />
             ) : (
                 <input {...inputParams} ref={getFieldNode} />
             )}
+            {renderLabel()}
             {!hideIcon && renderIconBlock()}
         </>
     );
@@ -388,37 +484,34 @@ const TextField: React.FC<TextFieldProps> = ({
                 {
                     disabled,
                     theme,
-                    valid: verification === Verification.VALID,
-                    error: verification === Verification.ERROR || isMaxLimitExceeded,
+                    valid: isValidVerification,
+                    error: isErrorVerification || isMaxLimitExceeded,
                     icon: !hideIcon && (!!verification || !!customIcon || type === 'password') && !textarea,
                     password: isPlaceholderShowed,
                 },
                 className,
             )}
         >
-            {label && (
-                <InputLabel dataAttrs={{ root: dataAttrs?.label }} htmlFor={id}>
-                    {label}
-                    {required && <span className={cn('require-mark')}>*</span>}
-                </InputLabel>
-            )}
-            <div className={cn('field-wrapper', { 'no-touch': !isTouch })}>{renderField()}</div>
-            <div className={cn('wrap')}>
-                {noticeText && (
-                    <div
-                        {...filterDataAttrs(dataAttrs?.notice)}
-                        className={cn('text', {
-                            error: verification === Verification.ERROR,
-                            success: verification === Verification.VALID,
-                        })}
-                    >
-                        {noticeText}
+            <div className={cn('field-wrapper', { textarea: textarea && textareaType })}>
+                {renderField()}
+                {textareaType === TextareaTypes.FLEXIBLE && (
+                    <div className={cn('resizer')} ref={resizerRef}>
+                        <ResizeIcon />
                     </div>
                 )}
+            </div>
+            <div className={cn('field-bottom-wrapper', { filled: !!currentNoticeText || !!symbolCounter })}>
+                <div
+                    className={cn('notice-text', { active: !!noticeText })}
+                    onTransitionEnd={handleNoticeTransitionEnd}
+                    {...filterDataAttrs(dataAttrs?.notice)}
+                >
+                    {currentNoticeText}
+                </div>
                 {symbolCounter && (
-                    <Paragraph size="small" hasMargin={false} className={cn('counter', { error: isMaxLimitExceeded })}>
+                    <span className={cn('counter', { error: isMaxLimitExceeded })}>
                         {`${currentSymbolCount}/${symbolCounter}`}
-                    </Paragraph>
+                    </span>
                 )}
             </div>
         </div>
@@ -437,6 +530,7 @@ TextField.propTypes = {
     inputMode: PropTypes.oneOf(['numeric', 'tel', 'decimal', 'email', 'url', 'search', 'none']),
     autoComplete: PropTypes.string,
     name: PropTypes.string,
+    hidePlaceholder: PropTypes.bool,
     placeholder: PropTypes.string,
     id: PropTypes.string,
     isControlled: PropTypes.bool,